Display Extended window on "double tap"
It is not really double tap but we cannot rely on timers. It display the window every two taps, and the counter is reset every minute or after displaying the extended window, so 2 taps will be required next time. If you tap once, wait 30 secs and tap again the extended window will display (will work on that later...)